+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade to Go 1.17
 - Upgrade github.com/giantswarm/microkit v0.2.2 to v1.0.0
 - Upgrade github.com/giantswarm/versionbundle v0.2.0 to v1.0.0
-- Upgrade github.com/giantswarm/microendpoint v0.2.0 to v0.3.0
+- Upgrade github.com/giantswarm/microendpoint v0.2.0 to v1.0.0
 - Upgrade github.com/giantswarm/microerror v0.3.0 to v0.4.0
 - Upgrade github.com/giantswarm/micrologger v0.5.0 to v0.6.0
 - Upgrade github.com/spf13/viper v1.9.0 to v1.10.0
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade k8s.io/apiextensions-apiserver v0.19.4 to v0.20.12
 - Upgrade sigs.k8s.io/controller-runtime v0.6.4 to v0.8.3
 - Upgrade k8s.io/client-go v0.19.4 to v0.20.12
+- Upgrade github.com/giantswarm/operatorkit v4.3.1 to v7.0.0
 
 ### Removed